override the exising sentinelhub_request.py with the new convenience request
