+++ sentinelhub/__init__.py
@@ -36,8 +36,7 @@ from .testing_utils import TestSentinelHub, TestCaseContainer
 from .download.sentinelhub_client import SentinelHubDownloadClient
 from .sentinelhub_session import SentinelHubSession
 
-from .sentinelhub_request import body, bounds, data, output, response
-from .sentinelhub_req import SentinelHubRequest
+from .sentinelhub_request import SentinelHubRequest
 from .download.cache import hash_request
 
 from .time_utils import parse_time_interval
